+++ test/platform/swift/unit_integration/mocks/MockLogging.swift
@@ -146,5 +146,7 @@ extension MockLogging: Logging {
         sleepMode = mode
     }
 
-    public func setFeatureFlagExposure(withName flag: String, variant: String?) {}
+    public func setFeatureFlagExposure(withName flag: String, variant: String) {}
+    
+    public func setFeatureFlagExposure(withName flag: String, variant: Bool) {}
 }